requirements.txt
fastapi
uvicorn
motor
dnspython
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
dotenv
certifi
APScheduler